refactor: rename action button
